IGNITE-14477: fix comments
