+++ modules/ducktests/tests/ignitetest/tests/rebalance/in_memory_test.py
@@ -20,7 +20,8 @@ from ducktape.mark import defaults
 
 from ignitetest.services.ignite import IgniteService
 from ignitetest.services.utils.ignite_configuration.discovery import from_ignite_cluster
-from ignitetest.tests.rebalance import preload_data, start_ignite, get_result, TriggerEvent, NUM_NODES
+from ignitetest.tests.rebalance.util import preload_data, start_ignite, get_result, TriggerEvent, NUM_NODES, \
+    await_rebalance_start
 from ignitetest.utils import cluster, ignite_versions
 from ignitetest.utils.ignite_test import IgniteTest
 from ignitetest.utils.version import DEV_BRANCH, LATEST
@@ -100,5 +101,9 @@ class RebalanceInMemoryTest(IgniteTest):
                                    ignites.config._replace(discovery_spi=from_ignite_cluster(ignites)), num_nodes=1)
             ignite.start()
             rebalance_nodes = ignite.nodes
+
+            await_rebalance_start(ignite)
+
+            ignite.await_rebalance()
 
         return get_result(rebalance_nodes, preload_time, cache_count, entry_count, entry_size)
